+++ src/dom/htmlelement.ts
@@ -102,8 +102,50 @@ export class HtmlElement extends DOMNode {
 		) as HtmlElement[];
 	}
 
+	/**
+	 * Find the first ancestor matching a selector.
+	 *
+	 * Implementation of DOM specification of Element.closest(selectors).
+	 */
+	public closest(selectors: string): HtmlElement {
+		let node: HtmlElement = this;
+		while (node) {
+			if (node.matches(selectors)) {
+				return node;
+			}
+			node = node.parent;
+		}
+		return null;
+	}
+
 	public is(tagName: string): boolean {
 		return (this.tagName && tagName === "*") || this.tagName === tagName;
+	}
+
+	/**
+	 * Match this element against given selectors. Returns true if any selector
+	 * matches.
+	 *
+	 * Implementation of DOM specification of Element.matches(selectors).
+	 */
+	public matches(selector: string): boolean {
+		/* find root element */
+		let root: HtmlElement = this;
+		while (root.parent) {
+			root = root.parent;
+		}
+
+		/* a bit slow implementation as it finds all candidates for the selector and
+		 * then tests if any of them are the current element. A better
+		 * implementation would be to walk the selector right-to-left and test
+		 * ancestors. */
+		for (const match of root.querySelectorAll(selector)) {
+			if (match.unique === this.unique) {
+				return true;
+			}
+		}
+
+		return false;
 	}
 
 	public setAttribute(