+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Upcoming release
 
+- add `requiredAncestors` metadata and validation to test elements with
+  additional requirements for the parent elements, such as `<area>` and
+  `<dd>`/`<dt>`.
+- add `HtmlElement.closest()` to locate a parent matching the given selector.
+- add `HtmlElement.matches()` to test if a selector matches the given element.
+- fix so multiple combinators can be used in selectors, e.g. `foo > bar > baz`
+  now works as expected.
+
 ## 1.1.2 (2019-06-18)
 
 - allow div to group elements under dl (fixes #44).